requirements.txt
flask
flask-socketio
python-dotenv
openai
pyautogui
keyboard
requests
python-socketio